+++ tests/test_schema.py
@@ -635,15 +635,29 @@ class TestSchemaInstantiation:
         from textwrap import dedent
 
         class TestSchema(schema.Schema):
-            foo = fields.String()
+            foo = fields.String(attr='foo_attr')
+            bar = fields.String()
 
         test_schema = TestSchema()
         expected = dedent(
-            '''def _dump_function(ser, obj):
+            '''\
+            def _dump_function(schema, obj):
                 return {
-                    "foo": obj.foo
+                    "foo": obj.foo_attr,
+                    "bar": obj.bar
                 }
             '''
         )
+        assert test_schema._get_dump_function_code() == expected
 
+        test_schema = TestSchema(ordered=True)
+        expected = dedent(
+            '''\
+            def _dump_function(schema, obj):
+                return OrderedDict([
+                    ("foo", obj.foo_attr),
+                    ("bar", obj.bar)
+                ])
+            '''
+        )
         assert test_schema._get_dump_function_code() == expected