+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Changelog
 
     While unreleased, the changelog of lima 0.3 is itself subject to change.
 
+- Support dumping of ``OrderedDict`` objects by providing ``ordered=True`` to
+  Schema constructor
+
 - Add New possibilities to specify a schema's fields:
 
     - Add support for ``__lima_args__['only']`` on schema definition